+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unpublished]
+## Changed
+- Changed method name from `get_price()` to `get_prices()` to reflect that the method get prices for 24 hours
+
 ## [0.3.1] - 2024-07-19
 ### Added
 - Implemented PartialOrd for HourlyPrice